Пока что всё работает
